Apply suggestions from review
